refactor(cache): 优化活动缓存键及缓存注解配置
- 统一使用 CacheKey.ACTIVITY_AREA 作为活动缓存区域
- 将deleteActivity方法的缓存策略由@CacheEvict改为@CachePut
- 修改updateActivity方法添加@CacheEvict注解，确保更新缓存一致性
- 统一调整各缓存注解的键表达式格式，增强规范性

feat(auth): 完善认证服务配置与登录逻辑

- 在AuthMainApplication中启用Feign客户端支持
- 添加Nacos、Feign、数据源、Redis及RabbitMQ等配置项
- 使用CacheKey构建Redis缓存键替代硬编码字符串
- 调整登录方法中的Redis缓存写入逻辑和Token生成逻辑
- 完善退出登录时的Redis缓存键生成和清理
- 增加AuthService单元测试覆盖登录和JWT功能

feat(user): 实现用户交互功能模块及相关接口

- 新增InteractionService接口及其实现类，实现用户交互记录的增删查
- 创建InteractionMapper实现数据库操作，支持删除交互记录
- 引入UserInteractionDto传递用户交互操作参数
- 实现用户点赞、收藏、关注、屏蔽、拉黑、禁言等操作的REST接口
- 每个接口通过RequestContext获取当前用户ID，调用InteractionService处理业务
- 各交互控制器均支持列表查询和新增、删除操作

chore(common): 提炼共用工具与配置模块为my-common-boot-starter

- 重命名common模块为my-common-boot-starter
- 迁移工具类IdUtil和JwtUtil，统一封装ID生成和JWT相关功能
- 新增CacheKey类统一管理缓存键常量及生成逻辑
- 新增Logical枚举支持权限逻辑判断
- 提取MybatisPlusConfig、RedisConfig和RabbitMQConfig自动装配配置类
- 添加CommonAutoConfiguration整合多个公共配置信息

fix(filter): 优化请求上下文Filter中请求头解析逻辑

- 在Auth服务中Filter中增加请求头Null及空串校验，避免异常
- 修改Resource服务中Filter直接转换Long，支持非空请求头
- 统一处理请求中的X-RequestId和X-UserId设置，保证请求上下文有效

refactor(gateway): 优化JWT过滤器中Token处理逻辑

- 优化JWTFilter中解析JWT Token逻辑，支持从Redis格式Key提取用户ID
- 引入CacheKey常量类替代硬编码字符串
- 增加日志输出，优化白名单处理的代码格式

build: 清理AI服务相关模块及配置文件

- 删除AI服务Main入口及相关模块配置文件
- 更新项目模块描述文件，移除gateway模块引用不完整部分
- 修正IDE配置文件编码路径，统一更改为my-common-boot-starter模块

test(auth): 新增登录及JWT相关单元测试

- 测试AuthService登录接口正确性
- 测试JwtUtil生成和解析JWT令牌正确性

style: 代码风格及细节改进

- 依照规范调整注解顺序及格式
- 去除无用导入，整理import顺序
- 优化注释描述，提升代码可读性

+++ services/auth-service/src/main/java/org/example/mapper/AuthorityMapper.java
@@ -14,5 +14,5 @@ public interface AuthorityMapper {
             "JOIN role_permissions rp ON ur.role_id = rp.role_id " +
             "JOIN permissions p ON rp.permission_id = p.id " +
             "WHERE u.id = #{id}")
-    List<String> selectUserPermissionById(Integer id);
+    List<String> selectUserPermissionById(Long id);
 }
